Create barGraph 2 functions
2 functions to handle the creation of the visual bar graphs.
the first one is for the start of the server
the 2nd one handles requests that changes when user submits a search
keyword. this function handles updates to the visuals and updates them
to the css accordingly. uses app.get code to handle this instead of
pasting the function in the request handler in app.post. Much cleaner