added in webb widget
this version of the api widget is 1.0,  the first version. So it may not
catch all the cities the user search. For example typing in hong kong
doesn;t work because there is a space. However typing in hongkong
together works. The developers of the widget need to fix that in the
next itereation of the version.